Add configuration to put microprofile scope in OpenMetrics tag
Add a `smallrye.metrics.usePrefixForScope` configuration property
(`true` by default) that specifies whether the microprofile scope
(`base`, `vendor` or `application`) must be prefixed to the OpenMetrics
name or added to the metric's tags (with the key `microprofile_scope`).

This allows applications that were already providing metrics in
Prometheus/OpenMetrics to switch to smallrye-metrics to expose their
metrics while keeping the same names for metrics.

Add OpenMetricsExporterTest#testMicroProfileScopeInTags
to test this configuration setting