FIX:
Including fix into this PR to expedite.

MetricRegistries logic moved to a SharedMetricRegistries class.
MetricRegistries was an application scoped CDI bean. This resulted in
the creation of a new MetricRegistries for each applicatoin deployed
(concurrently or sequentially). This resulted in multiple App, Base and
Vendor (+ custom) metric registries being registered to the Global
micrometer composite registry. MetricRegistries has now been turned to a
MetricRegistryProducer as all it does is produce the registry defined in
the RegistryType scope parameter.