FIX:
Include fix to expedite.

When registering to the global composite registry, the composite meter
that is created attempts to register to each registry in the composite
registry. If a registry denies the registration due to a use of a filter
then a noop meter is returned to the composite meter and is added  to
the list of registry that it tracks. When obtaining values from the
composite meter it will interrogate its registry x meter map for the
first entry. If the first registry denied registration and returned a
no-op, the value retrieval of the composite meter will be 0. This commit
fixes this by explicitly querying the exact prometheus meter registry
that the metric adapter belongs to to get the right values for the
Metric API call.

Additionally, The TimerAdapter and HistogrmAdapter have been fleshed
out. Remaining is the Snapshot related call, but that will be adressed
in another PR (as there will need to be a  PR for the MP Metrics
repository that needs to be approved first before work can be done).

Also some fixups in LegacyMetricRegistryAdapter. Remove unused method
and refactor remove method.